refactor: split project into binary and library creates
This way we can make integration tests using rust's tests/ directory
